refactor(client): P5c — agents/tasks handler domain (#142)
* refactor(client): P5c — agents/tasks/loops handlers to native_tools/agents_tasks.py

The deferred fifth domain (RFC R4), moved now that P7/P8 settled the loop
terrain. Verbatim host-based moves of all 15 handlers + _collect_agent_
result (~550 lines): delegate_task/list/cancel, start/stop/list loops,
spawn/send/list/kill/results/wait agents, spawn/collect loop agents. The
loop-coupled closures route through bot delegates (bot._run_loop_
iteration, bot._dispatch_loop_tool, bot._emit_lifecycle_event) — the
late-bound seams.

Declared migrations: (1) test_scheduled_workflow's 4 unbound
OdinBot._collect_agent_result(bot, ...) call sites now construct
AgentTaskTools(bot) — same real body, assertions untouched; (2) the
spawn-agent saver-wiring getsource assertion retargets to
AgentTaskTools._handle_spawn_agent with the bot. spelling (sec 8.4).
Normalizations (disclosed): _REQUIRED_FIELDS local renamed
required_fields; one truly-unused channel_id local dropped in
spawn_agent; ruff-format wrapping.

client.py: 2,290 -> 1,789 lines. Characterization 127 green, full suite
6040/4, agents_tasks.py ruff-clean.

* test: drop orphaned OdinBot import in retargeted getsource test (review tidy)
