+++ tests/test_trajectory_completeness.py
@@ -454,6 +454,8 @@ class TestAgentSaverWiring:
         # The omission hid for months because nothing asserted the call site.
         import inspect
 
-        from src.discord.client import OdinBot
-        src = inspect.getsource(OdinBot._handle_spawn_agent)
-        assert "trajectory_saver=self.agent_trajectory_saver" in src
+        # P5c: body moved to native_tools/agents_tasks.py (host-based)
+        from src.discord.native_tools.agents_tasks import AgentTaskTools
+
+        src = inspect.getsource(AgentTaskTools._handle_spawn_agent)
+        assert "trajectory_saver=bot.agent_trajectory_saver" in src
